add one unit test, move shared logic to directive
